# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      parameters:
        - name: withBranding
          in: query
          description: When true, include letterheadImage and signatureImage (heavy base64 fields). Defaults to false.
          schema:
            type: boolean
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client (also unlinks any allocated candidates)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /expense-categories:
    get:
      operationId: listExpenseCategories
      tags: [expenses]
      summary: List expense categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategory"
    post:
      operationId: createExpenseCategory
      tags: [expenses]
      summary: Create an expense category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseCategory"
        "409":
          description: A category with this name already exists
  /expense-categories/{id}:
    patch:
      operationId: updateExpenseCategory
      tags: [expenses]
      summary: Update an expense category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseCategory"
        "404":
          description: Not found
        "409":
          description: Another category with this name already exists
    delete:
      operationId: deleteExpenseCategory
      tags: [expenses]
      summary: Delete an expense category (refused if expenses still reference it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
        "409":
          description: Category is in use by one or more expenses
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Client:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ExpenseCategory:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ExpenseCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
    ExpenseCategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: ["string", "null"]
...
    Company:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        signatoryName:
          type: ["string", "null"]
        signatoryDesignation:
          type: ["string", "null"]
        letterheadImage:
          type: ["string", "null"]
          description: Base64 data URL (image/png or image/jpeg) shown at the top of generated LOA PDFs
        signatureImage:
          type: ["string", "null"]
          description: Base64 data URL (image/png or image/jpeg) inserted on the signature line of generated LOA PDFs
        createdAt:
          type: string
        updatedAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        registrationNumber:
          type: string
        signatoryName:
          type: string
        signatoryDesignation:
          type: string
        letterheadImage:
          type: string
        signatureImage:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        registrationNumber:
          type: string
        signatoryName:
          type: string
        signatoryDesignation:
          type: string
        letterheadImage:
          type: ["string", "null"]
        signatureImage:
          type: ["string", "null"]
...